// src/api/generated/model/createExpenseRequestCategory.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * TaxMate API
 * 프리랜서 세금 관리 서비스 API
 * OpenAPI spec version: 1.0.0
 */

export type CreateExpenseRequestCategory = typeof CreateExpenseRequestCategory[keyof typeof CreateExpenseRequestCategory];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateExpenseRequestCategory = {
  OFFICE_SUPPLY: 'OFFICE_SUPPLY',
  EQUIPMENT: 'EQUIPMENT',
  SOFTWARE: 'SOFTWARE',
  TRANSPORT: 'TRANSPORT',
  MEAL: 'MEAL',
  EDUCATION: 'EDUCATION',
  TELECOM: 'TELECOM',
  RENT: 'RENT',
  OUTSOURCING: 'OUTSOURCING',
  TAX: 'TAX',
  INSURANCE: 'INSURANCE',
  ETC: 'ETC',
} as const;